example/requirements.txt
# Your app requirements.
Django==1.9.5
psycopg2==2.6.1
django-dotenv
django-debug-toolbar

# Submit app
-e ../